Last minitalk fixes

+++ sources/client.c
@@ -6,7 +6,7 @@
 /*   By: maxperei <maxperei@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/17 15:29:48 by tulipe            #+#    #+#             */
-/*   Updated: 2022/04/21 14:30:14 by maxperei         ###   ########lyon.fr   */
+/*   Updated: 2022/05/11 10:57:27 by maxperei         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	send_signal(pid_t server_pid, char *str)
 {
-	int i;
-	int j;
+	int	i;
+	int	j;
 
 	i = 0;
 	while (str[i])
@@ -25,7 +25,7 @@ void	send_signal(pid_t server_pid, char *str)
 		{
 			kill(server_pid, SIGUSR1 + ((str[i] >> j) & 1));
 			j++;
-			usleep(100);
+			usleep(50);
 		}
 		i++;
 	}
@@ -33,14 +33,19 @@ void	send_signal(pid_t server_pid, char *str)
 	while (j < 8)
 	{
 		kill(server_pid, SIGUSR1);
-		usleep(100);
+		usleep(50);
 		j++;
 	}
 }
 
-int main(int argc, char **argv)
+int	main(int argc, char **argv)
 {
 	if (argc < 3)
 		return (1);
+	if (ft_atoi(argv[1]) == 0)
+	{
+		ft_printf("ERROR: bad pid\n");
+		return (0);
+	}
 	send_signal(ft_atoi(argv[1]), argv[2]);
 }
